Streamers are monitored and user is notified.
Twitch API is queried for currently hard-coded streamer list. If streamers are found then individual notifications are sent out.

Notifications are tracked and will redirect when clicked on, these are deleted after 1 minute to prevent any memory issues with the object accumulating too much. Should not be a concern on modern hardware though.

Authentication tokens are gathered from storage and query OAuth if required.